+++ src/pyrite/_services/camera_service/__init__.py
