+++ src/pyrite/_services/camera_service/camera_service.py
